+++ payment_adyen_alokai/graphql/__init__.py
@@ -1,0 +1,1 @@
+from . import adyenPayment
