+++ payment_adyen_alokai/graphql/adyenPayment.py
@@ -1,0 +1,251 @@
+import graphene
+from graphene.types import generic
+from graphql import GraphQLError
+from odoo import _
+from odoo.tools import format_amount
+from odoo.addons.graphql_alokai.graphql.registry import query_registry, mutation_registry
+
+from odoo.addons.payment import utils as payment_utils
+from odoo.addons.payment_adyen_alokai.const import CURRENCY_DECIMALS
+from odoo.addons.website_sale.controllers.payment import PaymentPortal
+from odoo.addons.payment_adyen.controllers.main import AdyenController
+from odoo.addons.payment_adyen_alokai.controllers.main import AdyenControllerInherit
+
+
+# -------------------------------- #
+#           Adyen Payment          #
+# -------------------------------- #
+
+class AdyenProviderInfoResult(graphene.ObjectType):
+    adyen_provider_info = generic.GenericScalar()
+
+
+class AdyenPaymentMethodsResult(graphene.ObjectType):
+    adyen_payment_methods = generic.GenericScalar()
+
+
+class AdyenTransactionResult(graphene.ObjectType):
+    transaction = generic.GenericScalar()
+
+
+class AdyenPaymentsResult(graphene.ObjectType):
+    adyen_payments = generic.GenericScalar()
+
+
+class AdyenPaymentDetailsResult(graphene.ObjectType):
+    adyen_payment_details = generic.GenericScalar()
+
+
+class AdyenProviderInfo(graphene.Mutation):
+    class Arguments:
+        provider_id = graphene.Int(required=True)
+
+    Output = AdyenProviderInfoResult
+
+    @staticmethod
+    def mutate(self, info, provider_id):
+        env = info.context["env"]
+        PaymentProvider = env['payment.provider'].sudo()
+        domain = [
+            ('id', '=', provider_id),
+            ('state', 'in', ['enabled', 'test']),
+        ]
+
+        payment_provider_id = PaymentProvider.search(domain, limit=1)
+        if not payment_provider_id:
+            raise GraphQLError(_('Payment Provider does not exist.'))
+
+        if not payment_provider_id.code == 'adyen':
+            raise GraphQLError(_('Payment Provider "Adyen" does not exist.'))
+
+        adyen_provider_info = {
+            'state': payment_provider_id.state,
+            'client_key': payment_provider_id.adyen_client_key
+        }
+
+        return AdyenProviderInfoResult(adyen_provider_info=adyen_provider_info)
+
+
+class AdyenPaymentMethods(graphene.Mutation):
+    class Arguments:
+        provider_id = graphene.Int(required=True)
+
+    Output = AdyenPaymentMethodsResult
+
+    @staticmethod
+    def mutate(self, info, provider_id):
+        env = info.context["env"]
+        PaymentProvider = env['payment.provider'].sudo()
+        website = env['website'].get_current_website()
+        order = website.sale_get_order()
+        domain = [
+            ('id', '=', provider_id),
+            ('state', 'in', ['enabled', 'test']),
+        ]
+
+        payment_provider_id = PaymentProvider.search(domain, limit=1)
+        if not payment_provider_id:
+            raise GraphQLError(_('Payment Provider does not exist.'))
+
+        if not payment_provider_id.code == 'adyen':
+            raise GraphQLError(_('Payment Provider "Adyen" does not exist.'))
+
+        adyen_payment_methods = AdyenController().adyen_payment_methods(
+            provider_id=payment_provider_id.id,
+            formatted_amount=format_amount(env, order.amount_total, order.currency_id),
+            partner_id=order.partner_id.id
+        )
+
+        return AdyenPaymentMethodsResult(adyen_payment_methods=adyen_payment_methods)
+
+
+class AdyenTransaction(graphene.Mutation):
+    class Arguments:
+        provider_id = graphene.Int(required=True)
+        tokenization_requested = graphene.Boolean(default_value=False)
+
+    Output = AdyenTransactionResult
+
+    @staticmethod
+    def mutate(self, info, provider_id, tokenization_requested):
+        env = info.context["env"]
+        PaymentProvider = env['payment.provider'].sudo()
+        PaymentTransaction = env['payment.transaction'].sudo()
+        website = env['website'].get_current_website()
+        order = website.sale_get_order()
+        domain = [
+            ('id', '=', provider_id),
+            ('state', 'in', ['enabled', 'test']),
+        ]
+
+        payment_provider_id = PaymentProvider.search(domain, limit=1)
+        payment_method_id = payment_provider_id.payment_method_ids[0].id if payment_provider_id.payment_method_ids else None
+        if not payment_provider_id:
+            raise GraphQLError(_('Payment Provider does not exist.'))
+
+        if not payment_provider_id.code == 'adyen':
+            raise GraphQLError(_('Payment Provider "Adyen" does not exist.'))
+
+        # Generate a new access token
+        access_token = payment_utils.generate_access_token(order.partner_id.id, order.amount_total, order.currency_id.id)
+        order.access_token = access_token
+
+        transaction = PaymentPortal().shop_payment_transaction(
+            order_id=order.id,
+            access_token=order.access_token,
+            provider_id=provider_id,
+            payment_method_id=payment_method_id,
+            token_id=None,
+            amount=order.amount_total,
+            flow='direct',
+            tokenization_requested=tokenization_requested,
+            landing_route='/shop/payment/validate'
+        )
+
+        transaction_id = PaymentTransaction.search([('reference', '=', transaction['reference'])], limit=1)
+
+        # Update the field created_on_alokai
+        transaction_id.created_on_alokai = True
+
+        return AdyenTransactionResult(transaction=transaction)
+
+
+class AdyenPayments(graphene.Mutation):
+    class Arguments:
+        provider_id = graphene.Int(required=True)
+        transaction_reference = graphene.String(required=True)
+        access_token = graphene.String(required=True)
+        payment_method = generic.GenericScalar(required=True, description='Return state.data.paymentMethod')
+        browser_info = generic.GenericScalar(required=True, description='Return state.data.browserInfo')
+
+    Output = AdyenPaymentsResult
+
+    @staticmethod
+    def mutate(self, info, provider_id, transaction_reference, access_token, payment_method, browser_info):
+        env = info.context["env"]
+        PaymentProvider = env['payment.provider'].sudo()
+        PaymentTransaction = env['payment.transaction'].sudo()
+        domain = [
+            ('id', '=', provider_id),
+            ('state', 'in', ['enabled', 'test']),
+        ]
+
+        payment_provider_id = PaymentProvider.search(domain, limit=1)
+        if not payment_provider_id:
+            raise GraphQLError(_('Payment Provider does not exist.'))
+
+        if not payment_provider_id.code == 'adyen':
+            raise GraphQLError(_('Payment Provider "Adyen" does not exist.'))
+
+        transaction = PaymentTransaction.search([('reference', '=', transaction_reference)], limit=1)
+        if not transaction:
+            raise GraphQLError(_('Payment transaction does not exist.'))
+
+        converted_amount = payment_utils.to_minor_currency_units(
+            transaction.amount,
+            transaction.currency_id,
+            arbitrary_decimal_number=CURRENCY_DECIMALS.get(transaction.currency_id.name, 2)
+        )
+
+        # Create Payment
+        adyen_payment = AdyenControllerInherit().adyen_payments(
+            provider_id=payment_provider_id.id,
+            reference=transaction.reference,
+            converted_amount=converted_amount,
+            currency_id=transaction.currency_id.id,
+            partner_id=transaction.partner_id.id,
+            payment_method=payment_method,
+            access_token=access_token,
+            browser_info=browser_info
+        )
+
+        return AdyenPaymentsResult(adyen_payments=adyen_payment)
+
+
+class AdyenPaymentDetails(graphene.Mutation):
+    class Arguments:
+        provider_id = graphene.Int(required=True)
+        transaction_reference = graphene.String(required=True)
+        payment_details = generic.GenericScalar(required=True, description='Return state.data')
+
+    Output = AdyenPaymentDetailsResult
+
+    @staticmethod
+    def mutate(self, info, provider_id, transaction_reference, payment_details):
+        env = info.context["env"]
+        PaymentProvider = env['payment.provider'].sudo()
+        PaymentTransaction = env['payment.transaction'].sudo()
+        domain = [
+            ('id', '=', provider_id),
+            ('state', 'in', ['enabled', 'test']),
+        ]
+
+        payment_provider_id = PaymentProvider.search(domain, limit=1)
+        if not payment_provider_id:
+            raise GraphQLError(_('Payment Provider does not exist.'))
+
+        if not payment_provider_id.code == 'adyen':
+            raise GraphQLError(_('Payment Provider "Adyen" does not exist.'))
+
+        transaction = PaymentTransaction.search([('reference', '=', transaction_reference)], limit=1)
+        if not transaction:
+            raise GraphQLError(_('Payment transaction does not exist.'))
+
+        # Submit the details
+        adyen_payment_details = AdyenController().adyen_payment_details(
+            provider_id=payment_provider_id.id,
+            reference=transaction.reference,
+            payment_details=payment_details
+        )
+
+        return AdyenPaymentDetailsResult(adyen_payment_details=adyen_payment_details)
+
+
+class AdyenPaymentMutation(graphene.ObjectType):
+    adyen_provider_info = AdyenProviderInfo.Field(description='Get Adyen Provider Info.')
+    adyen_payment_methods = AdyenPaymentMethods.Field(description='Get Adyen Payment Methods.')
+    adyen_transaction = AdyenTransaction.Field(description='Create Adyen Transaction')
+    adyen_payments = AdyenPayments.Field(description='Make Adyen Payment request.')
+    adyen_payment_details = AdyenPaymentDetails.Field(description='Submit the Adyen Payment Details.')
+
+mutation_registry.append(AdyenPaymentMutation)
